+++ src/database/account/models.py
@@ -1,8 +1,12 @@
-from sqlmodel import SQLModel, Field
+from sqlalchemy import Column, Time
+from sqlmodel import Field
+from decimal import Decimal
 from typing import Optional
-from datetime import datetime
-
+from datetime import datetime, time
+from enum import Enum
+from pydantic import field_validator
 from src.database.base import SQLModelLU
+
 
 class Account(SQLModelLU, table=True):
   __tablename__ = "account" # type: ignore
@@ -11,7 +15,7 @@ class Account(SQLModelLU, table=True):
   email: str
 
   # auth, ONE of these needs to be here
-  hashed_password: Optional[str] = None
+  hashed_password: Optional[str] = Field(default=None)
   gcp_user_id: Optional[str] = None
   
   # demo
@@ -27,3 +31,25 @@ class Account(SQLModelLU, table=True):
   admin_id: Optional[int] = Field(default=None, foreign_key="admin.id")
 
   created_at: Optional[datetime] = Field(default_factory=datetime.utcnow)
+  
+from enum import Enum
+
+class Weekday(str, Enum):
+   MONDAY = "monday"
+   TUESDAY = "tuesday"
+   WEDNESDAY = "wednesday"
+   THURSDAY = "thursday"
+   FRIDAY = "friday"
+   SATURDAY = "saturday"
+   SUNDAY = "sunday"
+
+class Availability(SQLModelLU, table=True):
+    __tablename__ = "availability"  # type: ignore
+
+    id: Optional[int] = Field(default=None, primary_key=True)
+    weekday: Weekday
+    start_time: time = Field(sa_type=Time(timezone=True), nullable=False)
+    end_time: time = Field(sa_type=Time(timezone=True), nullable=False)
+    max_time_commitment_seconds: Optional[Decimal] = Field(default=None, max_digits=8, decimal_places=2)
+    client_availability_id: Optional[int] = Field(default=None, foreign_key="client_availability.id")
+    coach_availability_id: Optional[int] = Field(default=None, foreign_key="coach_availability.id")