+++ src/database/client/models.py
@@ -7,15 +7,14 @@ from src.database.base import SQLModelLU
 class Client(SQLModelLU, table=True):
   __tablename__ = "client"  # type: ignore
   id: Optional[int] = Field(default=None, primary_key=True)
-  payment_information_id : int = Field(foreign_key="payment_information.id")
-  client_availability_id : int = Field(default=None, foreign_key="client_availability.id") #TODO should not be NONE, for MVP
+  payment_information_id : Optional[int] = Field(default=None, foreign_key="payment_information.id")
+  client_availability_id : Optional[int] = Field(default=None, foreign_key="client_availability.id")
 
 class ClientAvailability(SQLModelLU, table=True):
   __tablename__ = "client_availability"  # type: ignore
   id : Optional[int] = Field(default=None, primary_key=True)
-  id_weekly : bool
 
-class FitnessGoalEnum(Enum):
+class FitnessGoalEnum(str, Enum):
   WEIGHT_LOSS = "weight loss"
   MAINTENENCE = "maintenence"
   MUSCLE_GAIN = "muscle gain"